refactored file upload to app container
