+++ src/containers/App/constants.js
@@ -10,9 +10,17 @@
  */
 
 export const DEFAULT_LOCALE = 'nl';
+
 export const AUTHENTICATE_USER = 'sia/App/AUTHENTICATE_USER';
 export const AUTHORIZE_USER = 'sia/App/AUTHORIZE_USER';
+
 export const SHOW_GLOBAL_ERROR = 'sia/App/SHOW_GLOBAL_ERROR';
 export const RESET_GLOBAL_ERROR = 'sia/App/RESET_GLOBAL_ERROR';
+
 export const LOGIN = 'sia/App/LOGIN';
 export const LOGOUT = 'sia/App/LOGOUT';
+
+export const UPLOAD_REQUEST = 'src/App/UPLOAD_REQUEST';
+export const UPLOAD_PROGRESS = 'src/App/UPLOAD_PROGRESS';
+export const UPLOAD_SUCCESS = 'src/App/UPLOAD_SUCCESS';
+export const UPLOAD_FAILURE = 'src/App/UPLOAD_FAILURE';
