+++ src/containers/App/reducer.js
@@ -12,7 +12,15 @@
 
 import { fromJS } from 'immutable';
 
-import { AUTHORIZE_USER, SHOW_GLOBAL_ERROR, RESET_GLOBAL_ERROR } from './constants';
+import {
+  AUTHORIZE_USER,
+  SHOW_GLOBAL_ERROR,
+  RESET_GLOBAL_ERROR,
+  UPLOAD_REQUEST,
+  UPLOAD_PROGRESS,
+  UPLOAD_SUCCESS,
+  UPLOAD_FAILURE
+} from './constants';
 
 // The initial state of the App
 export const initialState = fromJS({
@@ -39,6 +47,18 @@ function appReducer(state = initialState, action) {
         .set('error', false)
         .set('errorMessage', '')
         .set('loading', false);
+
+    case UPLOAD_REQUEST:
+      return state;
+
+    case UPLOAD_PROGRESS:
+      return state;
+
+    case UPLOAD_SUCCESS:
+      return state;
+
+    case UPLOAD_FAILURE:
+      return state;
 
     default:
       return state;
