+++ src/signals/incident/containers/IncidentContainer/constants.js
@@ -13,8 +13,3 @@ export const CREATE_INCIDENT_ERROR = 'src/IncidentContainer/CREATE_INCIDENT_ERRO
 export const GET_CLASSIFICATION = 'src/IncidentContainer/GET_CLASSIFICATION';
 export const GET_CLASSIFICATION_SUCCESS = 'src/IncidentContainer/GET_CLASSIFICATION_SUCCESS';
 export const GET_CLASSIFICATION_ERROR = 'src/IncidentContainer/GET_CLASSIFICATION_ERROR';
-
-export const UPLOAD_REQUEST = 'src/IncidentContainer/UPLOAD_REQUEST';
-export const UPLOAD_PROGRESS = 'src/IncidentContainer/UPLOAD_PROGRESS';
-export const UPLOAD_SUCCESS = 'src/IncidentContainer/UPLOAD_SUCCESS';
-export const UPLOAD_FAILURE = 'src/IncidentContainer/UPLOAD_FAILURE';
